Rename Textarea to LinkingTextarea
