+++ package.json
@@ -34,7 +34,6 @@
     "typescript": "^5.3.3"
   },
   "dependencies": {
-    "linkify-string": "^4.1.3",
-    "rollup-plugin-import-css": "^3.4.0"
+    "linkify-string": "^4.1.3"
   }
 }
